trading_bot/requirements.txt
# Drift Protocol Trading Bot Dependencies
# Python 3.13 compatible versions

# Core dependencies
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3

# Date/time handling
pendulum==3.0.0

# Networking
websockets==13.0

# Fix for Python 3.13 compatibility
cffi>=1.15.1            # Updated for Python 3.13 support

# Solana and Drift Protocol (install without strict dependency resolution)
driftpy==0.8.76
anchorpy==0.21.0
solana==0.36.6
ta==0.10.2               # Technical analysis library for MACD strategy
# pandas-ta==0.3.14b0    # Alternative technical indicators (optional)
# ta-lib                 # Alternative technical analysis library
# Development/Optional
# matplotlib==3.8.0      # For plotting (install separately if needed)

# Visualization (optional - install separately if needed)
# matplotlib>=3.7.0
# mplfinance>=0.12.7a0
streamlit